个人修改
1.主要解决snprintf在VS2015中的支持
2.解决项目在VS平台下出现的告警信息
3.解决在锁定变量时示将销变量的值置1

+++ pool/core/cpool_core_gc.h
@@ -36,7 +36,7 @@ static inline void cpool_core_GC_leavel(cpool_core_t *core, thread_t *self)
 
 static inline long cpool_core_GC_gettimeol(cpool_core_t *core, thread_t *self)
 {
-	unsigned us;
+	long us;
 
 	assert (core->us_gc_left_timeo >= 0);
